tools: Added install_artifacts_rootfs script
Added a new script, 'install_artifacts_rootfs.sh'. This script
takes a rootfs image and a directory with linux userlevel
artifacts to overlay ontop of the root filesystem. The script can
set up the rootfs image such that the artifacts are mounted as an
an overlayed filesystem (overlayfs) or are rather installed onto the
physical root file system.