CMake: Added CMake to camkes-linux-artifacts
Added CMake support to the camkes-linux-artifacts directory.
Since the programs are to be compiled for a different platform
(32-bit Linux), the CMake build for the camkes-linux-artifacts
directory is to be treated as an External project. The CMake
project has its own helpers for downloading the Linux sources and
defining Linux modules with Kbuild.