tools: Python tool to update dtb initrd values
Implemented the python tool 'update_dtb_initrd.py'. Given a
dtb and rootfs/initrd image, the tool will walk the dtb nodes
and update the initrd-start and initrd-end fields based on
the rootfs file size. These values need to be accurate in
the dtb in order to properly boot into the rootfs image.