connector-modules: Re-implemented dataport module
Re-implemented the dataport module to present itself as a PCI
driver rather than a character device. This aligns with the cross
vm changes in camkes-vm where shared dataports are now advertised
as a PCI device (where dataports are memory bars). This now parses
our Dataports PCI device and lifts the memory to Linux processes
in the form of userspace IO (UIO). This greatly simplifies the
process of creating shared dataports as it removes the hypercall
interface, making it more portable to other platforms that support
PCI.