connector-modules: Added PCI connector module
Re-implemented crossvm connections such that a connection
is now advertised as a PCI device. The connection driver aligns
with the camkes-vm changes where shared dataports and camkes
events are now advertised as PCI bars on a PCI device. Crossvm
connection devices will have an event bar and a series
of dataports which the module lifts to Linux processes in the
form of userspace IO (UIO). This greatly simplifies the
process of creating crossvm connections as it removes the
hypercall interface, consolidates all functionality to one driver
and makes it more portable to other platforms that support PCI.